move mock row model into separate file
